+++ Gw2Sharp/WebApi/V2/Models/Items/ItemKey.cs
@@ -1,0 +1,9 @@
+namespace Gw2Sharp.WebApi.V2.Models
+{
+    /// <summary>
+    /// Represents a key item.
+    /// </summary>
+    public class ItemKey : Item
+    {
+    }
+}